feat(product): 新增商品 gRPC 服务及相关 Proto 定义
- 新增 product_proto.proto，定义商品相关的 gRPC 请求和响应消息及服务接口
- 实现 ProductGrpcService，完成 CreateProduct、GetProduct 和 ReduceStock 接口方法
- 修改产品服务相关代码，统一使用 int64 类型进行库存和 ID 管理
- 重命名请求和响应实体包名为 request 和 response，规范代码结构
- 自动生成 Proto 相关的 Go 代码以支持 gRPC 服务调用
- 在 mian.go 添加对数据库连接的打印，方便调试验证数据库初始化情况

+++ app/product/internal/repository/product_repository.go
@@ -26,7 +26,7 @@ func (pr *ProductRepository) GetProduct(ctx context.Context, id int) (product *m
 	return product, err
 }
 
-func (pr *ProductRepository) ReduceStock(ctx context.Context, id, count int) (bool, error) {
+func (pr *ProductRepository) ReduceStock(ctx context.Context, id, count int64) (bool, error) {
 	tx := pr.db.Where("id = ?", id).Update("stock", gorm.Expr("stock - ?", count))
 	return tx.Error == nil, tx.Error
 }
